requirements.txt
argh==0.26.2
boto3==1.7.38
botocore==1.10.38
cachetools==2.1.0
certifi==2018.4.16
chardet==3.0.4
click==6.7
DateTime==4.2
delegator.py==0.1.0
dill==0.2.7.1
docutils==0.14
Flask==1.0.2
future==0.16.0
gapic-google-cloud-datastore-v1==0.15.3
gapic-google-cloud-error-reporting-v1beta1==0.15.3
gapic-google-cloud-logging-v2==0.91.3
gcloud==0.18.3
google-api-core==0.1.4
google-api-python-client==1.7.3
google-auth==1.5.0
google-auth-httplib2==0.0.3
google-cloud==0.33.1
google-cloud-bigquery==0.28.0
google-cloud-bigquery-datatransfer==0.1.1
google-cloud-bigtable==0.28.1
google-cloud-container==0.1.1
google-cloud-core==0.28.1
google-cloud-datastore==1.4.0
google-cloud-dns==0.28.0
google-cloud-error-reporting==0.28.0
google-cloud-firestore==0.28.0
google-cloud-language==1.0.2
google-cloud-logging==1.4.0
google-cloud-monitoring==0.28.1
google-cloud-pubsub==0.30.1
google-cloud-resource-manager==0.28.1
google-cloud-runtimeconfig==0.28.1
google-cloud-spanner==0.29.0
google-cloud-speech==0.30.0
google-cloud-storage==1.6.0
google-cloud-trace==0.17.0
google-cloud-translate==1.3.1
google-cloud-videointelligence==1.0.1
google-cloud-vision==0.29.0
google-gax==0.15.16
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
gpiozero==1.4.1
grpc-google-iam-v1==0.11.4
grpcio==1.12.1
httplib2==0.11.3
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
MarkupSafe==1.0
oauth2client==3.0.0
pathtools==0.1.2
pexpect==4.6.0
Pillow==5.1.0
ply==3.8
proto-google-cloud-datastore-v1==0.90.4
proto-google-cloud-error-reporting-v1beta1==0.15.3
proto-google-cloud-logging-v2==0.91.3
protobuf==3.6.0
protobuf3==0.2.1
protobuf3-to-dict==0.1.5
psutil==5.4.6
ptyprocess==0.5.2
pyasn1==0.4.3
pyasn1-modules==0.2.1
pybase64==0.4.0
python-dateutil==2.7.3
pytz==2018.4
PyYAML==3.12
requests==2.19.1
rsa==3.4.2
s3transfer==0.1.13
SE==2.3
six==1.11.0
uritemplate==3.0.0
urllib3==1.23
watchdog==0.8.3
Werkzeug==0.14.1
zope.interface==4.5.0
